Move aggregator list into separate view
