+++ client/src/lib/Sources/Overview.svelte
@@ -17,15 +17,8 @@
   import { tdClass } from "$lib/Table/defaults";
   import { request } from "$lib/request";
   import { onDestroy, onMount } from "svelte";
-  import { type CSAFAggregator } from "$lib/aggregatorTypes";
   import CustomTable from "$lib/Table/CustomTable.svelte";
-  import {
-    type Source,
-    type Aggregator,
-    fetchSources,
-    fetchAggregators,
-    fetchAggregatorData
-  } from "$lib/Sources/source";
+  import { type Source, fetchSources } from "$lib/Sources/source";
   import { appStore } from "$lib/store";
   import { DAY_MS } from "$lib/time";
   import ImportStats from "$lib/Statistics/ImportStats.svelte";
@@ -33,15 +26,10 @@
 
   let messageError: ErrorDetails | null;
   let sourcesError: ErrorDetails | null;
-  let aggregatorError: ErrorDetails | null;
 
   let loadingSources: boolean = false;
-  let loadingAggregators: boolean = false;
 
   let sources: Source[] = [];
-  let aggregators: Aggregator[] = [];
-  let aggregatorData = new Map<number, [string, string][]>();
-
   async function getMessage() {
     const response = await request("api/sources/message", "GET");
     if (response.ok) {
@@ -70,40 +58,9 @@
     }
   };
 
-  const getAggregators = async () => {
-    loadingAggregators = true;
-    const result = await fetchAggregators();
-    loadingAggregators = false;
-    if (result.ok) {
-      aggregators = result.value;
-    } else {
-      aggregatorError = result.error;
-    }
-  };
-
-  const parseAggregatorData = (data: CSAFAggregator): [string, string][] => {
-    return data.csaf_providers.map((i: any) => [i.metadata.publisher.name, i.metadata.url]);
-  };
-
-  const toggleAggregatorView = async (aggregator: Aggregator) => {
-    if (aggregatorData.get(aggregator.id ?? -1)) {
-      aggregatorData.delete(aggregator.id ?? -1);
-      aggregatorData = aggregatorData;
-      return;
-    }
-    const resp = await fetchAggregatorData(aggregator.url);
-    if (resp.ok) {
-      aggregatorData.set(aggregator.id ?? -1, parseAggregatorData(resp.value.aggregator));
-      aggregatorData = aggregatorData;
-    } else {
-      aggregatorError = resp.error;
-    }
-  };
-
   onMount(async () => {
     if (appStore.isEditor() || appStore.isSourceManager()) {
       await getSources();
-      await getAggregators();
     }
   });
 
@@ -193,59 +150,6 @@
         <ErrorMessage error={sourcesError}></ErrorMessage>
       </div>
     </CustomTable>
-    <CustomTable
-      title="Aggregator"
-      headers={[
-        {
-          label: "Name",
-          attribute: "name"
-        },
-        {
-          label: "URL",
-          attribute: "url"
-        }
-      ]}
-    >
-      {#each aggregators as aggregator, index (index)}
-        <tr
-          on:click={async () => {
-            if (appStore.isSourceManager()) {
-              await toggleAggregatorView(aggregator);
-            }
-          }}
-          on:blur={() => {}}
-          on:focus={() => {}}
-          class={appStore.isSourceManager() ? "cursor-pointer" : ""}
-        >
-          <TableBodyCell {tdClass}>{aggregator.name}</TableBodyCell>
-          <TableBodyCell {tdClass}>{aggregator.url}</TableBodyCell>
-        </tr>
-        {@const list = aggregatorData.get(aggregator.id ?? -1) ?? []}
-        {#each list as entry}
-          <tr class="bg-slate-200">
-            <TableBodyCell {tdClass}>{entry[0]}</TableBodyCell>
-            <TableBodyCell {tdClass}>{entry[1]}</TableBodyCell>
-          </tr>
-        {/each}
-      {/each}
-      <div slot="bottom">
-        <div
-          class:invisible={!loadingAggregators}
-          class={loadingAggregators ? "loadingFadeIn" : ""}
-          class:mb-4={true}
-        >
-          Loading ...
-          <Spinner color="gray" size="4"></Spinner>
-        </div>
-        {#if appStore.isSourceManager()}
-          <Button href="/#/aggregator/new" class="mb-2" color="primary" size="xs">
-            <i class="bx bx-plus"></i>
-            <span>Add aggregator</span>
-          </Button>
-        {/if}
-        <ErrorMessage error={aggregatorError}></ErrorMessage>
-      </div>
-    </CustomTable>
   {/if}
   {#await getMessage() then resp}
     {#if resp.message}
@@ -255,6 +159,12 @@
   <ErrorMessage error={sourcesError}></ErrorMessage>
   <ErrorMessage error={messageError}></ErrorMessage>
 
+  <br />
+  {#if appStore.isSourceManager()}
+    <Button href="/#/aggregator" class="my-2" color="primary" size="xs">
+      <span>Aggregator List</span>
+    </Button>
+  {/if}
   <br />
   {#if appStore.isImporter()}
     <Button href="/#/sources/upload" class="my-2" color="primary" size="xs">
